+++ dev-packages/application-manager/src/generator/backend-generator.ts
@@ -78,7 +78,7 @@ const { BackendApplicationConfigProvider } = require('@theia/core/lib/node/backe
 BackendApplicationConfigProvider.set(${this.prettyStringify(this.pck.props.backend.config)});
 
 const serverPath = require('path').resolve(__dirname, 'server');
-const address = require('@theia/core/lib/node/cluster/main').default(serverPath);
+const address = require('@theia/core/lib/node/main').default(serverPath);
 address.then(function (address) {
     if (process && process.send) {
         process.send(address.port.toString());